frontend/requirements.txt
streamlit==1.31.0
psycopg2-binary==2.9.9
pandas==2.2.0
sqlalchemy==2.0.25
